# docs/swagger.yaml
basePath: /api
definitions:
  models.CreateEmployeeRequest:
    properties:
      city:
        example: Noida
        type: string
      country:
        example: India
        type: string
      department:
        example: Engineering
        type: string
      designation:
        example: iOS Engineer
        type: string
      email:
        example: nitin@test.com
        type: string
      img_url:
        example: https://image.com/profile.jpg
        type: string
      joining_date:
        example: "2024-01-01"
        type: string
      mobiles:
        items:
          $ref: '#/definitions/models.MobileRequest'
        type: array
      name:
        example: Nitin Jain
        type: string
    required:
    - email
    - name
    type: object
  models.Employee:
    properties:
      city:
        type: string
      country:
        type: string
      department:
        type: string
      designation:
        type: string
      email:
        type: string
      id:
        type: string
      img_url:
        type: string
      is_active:
        type: boolean
      joining_date:
        type: string
      mobiles:
        items:
          $ref: '#/definitions/models.Mobile'
        type: array
      name:
        type: string
    required:
    - email
    type: object
  models.EmployeeDetailResponse:
    properties:
      data:
        $ref: '#/definitions/models.Employee'
      message:
        type: string
      status:
        type: string
    type: object
  models.EmployeeFilters:
    properties:
      departments:
        items:
          type: string
        type: array
      designations:
        items:
          type: string
        type: array
      mobile_types:
        items:
          $ref: '#/definitions/models.FilterOption'
        type: array
      statuses:
        items:
          $ref: '#/definitions/models.FilterOption'
        type: array
    type: object
  models.EmployeeFiltersResponse:
    properties:
      data:
        $ref: '#/definitions/models.EmployeeFilters'
      message:
        type: string
      status:
        type: string
    type: object
  models.EmployeeListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Employee'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/models.Meta'
      status:
        type: string
    type: object
  models.FilterOption:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  models.Meta:
    properties:
      has_next_page:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
  models.Mobile:
    properties:
      employee_id:
        type: string
      id:
        type: string
      number:
        type: string
      type:
        description: home, office, other
        type: string
    type: object
  models.MobileRequest:
    properties:
      number:
        example: "9876543210"
        type: string
      type:
        example: home
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Employee management service
  title: Employee API
  version: "1.0"
paths:
  /api/employees/{id}:
    patch:
      consumes:
      - application/json
      description: Update employee details by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Employee payload
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: Employee updated successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Missing ID or invalid request body
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Update an existing employee
      tags:
      - employees
  /employees:
    get:
      consumes:
      - application/json
      description: Fetch employees with pagination, search, and filters
      parameters:
      - description: Number of records to fetch
        example: 10
        in: query
        name: limit
        type: integer
      - description: Number of records to skip
        example: 0
        in: query
        name: offset
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Comma-separated designations
        in: query
        name: designation
        type: string
      - description: Comma-separated departments
        in: query
        name: department
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmployeeListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Get employees with filters
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Create a new employee with the provided details
      parameters:
      - description: Employee payload
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Employee created successfully
          schema:
            $ref: '#/definitions/models.EmployeeDetailResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Create a new employee
      tags:
      - employees
  /employees/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an employee by its ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee deleted successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Missing employee ID
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Delete an employee
      tags:
      - employees
    get:
      consumes:
      - application/json
      description: Fetch a single employee by its unique ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmployeeDetailResponse'
        "400":
          description: Missing or invalid ID
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Get employee by ID
      tags:
      - employees
  /employees/filters:
    get:
      consumes:
      - application/json
      description: Fetch available filter values like designation, department, status,
        etc.
      produces:
      - application/json
      responses:
        "200":
          description: Filter data fetched successfully
          schema:
            $ref: '#/definitions/models.EmployeeFiltersResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: Get employee filter options
      tags:
      - employees
swagger: "2.0"